Improvements to WMM implementation and added GPS Nav Packet

+++ external/wmm/src/GeomagnetismHeader.h
@@ -66,11 +66,11 @@
 #define ATanH(x)	    (0.5 * log((1 + x) / (1 - x)))
 
 #ifndef TRUE
-#define TRUE            ((int)1)
+#define TRUE            1
 #endif
 
 #ifndef FALSE
-#define FALSE           ((int)0)
+#define FALSE           0
 #endif
